additional checks for messages

+++ src/slashCommandManager.ts
@@ -3,6 +3,7 @@ import {
   CommandInteraction,
   Guild,
   GuildMember,
+  GuildTextBasedChannel,
   InteractionReplyOptions,
   InteractionUpdateOptions,
   Message,
@@ -83,18 +84,10 @@ export default class {
       }
     }
 
-    // logger.info('Loading commands overwrites...')
-    // const overwritesFiles = await globPromise('**/dist/slashOverwrites/*.js')
-    //
-    // for (const file of overwritesFiles) {
-    //   const module = await import(`../${file}`)
-    //   const command: CommandType = module.default
-    //
-    //   this.client.slashOverwrites.set(command.name, command)
-    // }
-
     this.client.on('interactionCreate', async (interaction) => {
       //console.log(interaction.options.data)
+      if (!interaction.inGuild()) return
+
       if (interaction.isCommand() || interaction.isButton())
         this.executeSlash(interaction).catch((err) =>
           logger.error(
@@ -118,15 +111,6 @@ export default class {
       //   })
       // }
     })
-
-    // обычные команды с префиксом
-    // this.client.on('messageCreate', async message => {
-    //   //console.log(message)
-    //   // this.executePrefix(message).catch((err) => logger.error({
-    //   //   shard_id: this.client.cluster.id,
-    //   //   err
-    //   // }, 'executePrefix'))
-    // })
   }
 
   /*
@@ -136,7 +120,7 @@ export default class {
     const guild = interaction.guild as Guild
     const user = interaction.member?.user as User
     const member = interaction.member as GuildMember
-    const text = interaction.channel as TextBasedChannel
+    const text = interaction.channel as GuildTextBasedChannel
     const voice = member?.voice?.channel
 
     const meta: Meta = {
@@ -145,38 +129,49 @@ export default class {
     }
 
     const respond = async (data: MessageOptions | InteractionReplyOptions, timeout?: number): Promise<void> => {
-      if (interaction.deferred)
+      if (interaction.deferred) {
         await interaction.editReply(data).catch((err) => logger.error({ err, ...meta }, "Can't edit reply"))
-      else
-        await interaction
-          .reply(data as InteractionReplyOptions)
-          .catch((err) => logger.error({ err, ...meta }, "Can't send reply"))
+        return
+      }
 
-      if (timeout)
-        setTimeout(async () => {
-          await interaction.deleteReply().catch((err) => logger.error({ err, ...meta }, 'Error deleting reply'))
-        }, timeout)
+      if (interaction.isRepliable()) {
+        try {
+          await interaction.reply(data as InteractionReplyOptions)
+        } catch (err) {
+          logger.error({ err, ...meta }, "Can't send reply")
+        }
+
+        if (timeout)
+          setTimeout(async () => {
+            try {
+              await interaction.deleteReply()
+            } catch (err) {
+              logger.error({ err, ...meta }, 'Error deleting reply')
+            }
+          }, timeout)
+      }
     }
 
     const send = async (data: MessageOptions, timeout?: number): Promise<void> => {
-      if (text.type !== 'GUILD_TEXT') return
-
       if (!text.permissionsFor(this.client.user as User)?.has(Permissions.FLAGS.SEND_MESSAGES)) return
 
-      const message = await text.send(data).catch((err) => logger.error({ err, ...meta }, "Can't send message"))
+      try {
+        const message = await text.send(data)
 
-      if (timeout)
-        setTimeout(async () => {
-          if (!message) return
+        if (timeout) {
+          setTimeout(async () => {
+            if (!message.channel.isText()) return
 
-          if (message.channel.type !== 'GUILD_TEXT') return
-
-          try {
-            await message.delete()
-          } catch (err) {
-            logger.error({ err, ...meta }, "Can't delete message")
-          }
-        }, timeout)
+            try {
+              if (message.deletable) await message.delete()
+            } catch (err) {
+              logger.error({ err, ...meta }, "Can't delete message")
+            }
+          }, timeout)
+        }
+      } catch (err) {
+        logger.error({ err, ...meta }, "Can't send message")
+      }
     }
 
     if (interaction.isCommand()) {
@@ -326,147 +321,6 @@ export default class {
 
   // async executeAutocomplete(interaction: AutocompleteInteraction) {
   //   if (interaction.commandName === 'play') {
-  //   }
-  // }
-
-  /*
-   * Обычные команды
-   */
-  // async executePrefix(message: Message): Promise<void> {
-  //   if (message.channel.type != 'GUILD_TEXT' || message.author.bot) return
-  //   if (!message.channel.permissionsFor(message.client.user as User)?.has(Permissions.FLAGS.SEND_MESSAGES)) return
-  //
-  //   const prefix = await this.client.db.getPrefix(message.guild?.id.toString())
-  //
-  //   if (!message.content.startsWith(prefix)) return
-  //
-  //   const args = message.content.slice(prefix.length).split(/ +/)
-  //
-  //   // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  //   // @ts-ignore
-  //   const commandName = args.shift().toLowerCase()
-  //
-  //   if (this.client.commands.has(commandName)) {
-  //     const guild = message.guild as Guild
-  //     const member = message.member as GuildMember
-  //     const channel = message.channel as TextBasedChannel
-  //
-  //     const meta = {
-  //       shard_id: this.client.cluster.id,
-  //       guild_id: guild.id
-  //     }
-  //
-  //     logger.info({ args, ...meta }, `Executed command ${commandName} with arguments`)
-  //
-  //     const respond = async (data: MessageOptions | InteractionReplyOptions, timeout?: number): Promise<void> => {
-  //       try {
-  //         const message = await channel.send(data as MessageOptions)
-  //
-  //         if (timeout)
-  //           setTimeout(() => {
-  //             if (message) {
-  //               message.delete().catch(err => logger.error({ err }, 'Can\'t delete message'))
-  //             }
-  //           }, timeout)
-  //       } catch (err) {
-  //         logger.error({ err }, 'Can\'t delete message')
-  //       }
-  //     }
-  //
-  //     const send = respond
-  //
-  //     const command = this.client.commands.get(commandName)
-  //
-  //     // проверка на админа
-  //     if (command?.adminOnly) {
-  //       if (!member.permissions.has(Permissions.FLAGS.MANAGE_GUILD)) {
-  //         await respond({
-  //           embeds: [Utils.generateErrorMessage('Эту команду могут выполнять только пользователи с правом `Управление сервером`.')]
-  //         })
-  //         return
-  //       }
-  //     }
-  //
-  //     // проверка на dj роль
-  //     if (await this.client.db.getAccessRoleEnabled(guild.id)) {
-  //       const djRole = await this.client.db.getAccessRole(guild.id)
-  //
-  //       if (!member.permissions.has(Permissions.FLAGS.MANAGE_GUILD) && !member.roles.cache.some(role => role.name === djRole)) {
-  //         await respond({
-  //           embeds: [Utils.generateErrorMessage(`Сейчас включен DJ режим, и вы не можете выполнять команды, так как у вас нет роли \`${djRole}\`.`)]
-  //         })
-  //         return
-  //       }
-  //     }
-  //
-  //     // проверка на наличие премиума
-  //     if (command?.premium && !await this.client.db.checkPremium(guild.id)) {
-  //       await respond({
-  //         embeds: [Utils.generateErrorMessage('Для выполнения этой команды требуется **Премиум**! Подробности: `/donate`')]
-  //       })
-  //       return
-  //     }
-  //
-  //     // проверка на наличие капчи
-  //     if (command?.name === 'play' || command?.name === 'search') {
-  //       const captcha = this.client.captcha.get(guild.id)
-  //
-  //       if (captcha) {
-  //         const embed = {
-  //           description: 'Ошибка! Ожидается команда, для которой не введена капча. Введите команду `/captcha`, а после код с картинки.' +
-  //             `Если картинки не видно, перейдите по [ссылке](${captcha.url})`,
-  //           color: 0x5181b8,
-  //           image: {
-  //             url: captcha.url + Utils.generateRandomCaptchaString()
-  //           }
-  //         }
-  //
-  //         await respond({ embeds: [embed] })
-  //         return
-  //       }
-  //     }
-  //
-  //     const realCommandName = command?.name as string
-  //
-  //     //проверка кулдауна
-  //     if (!this.client.cooldowns.has(realCommandName)) {
-  //       this.client.cooldowns.set(realCommandName, new Collection())
-  //     }
-  //
-  //     const now = Date.now()
-  //     const timestamps = this.client.cooldowns.get(realCommandName)
-  //     const cooldownAmount = (command?.cooldown || 3) * 1000
-  //
-  //     if (timestamps.has(member.user.id)) {
-  //       const expirationTime = timestamps.get(member.id) + cooldownAmount
-  //
-  //       if (now < expirationTime) {
-  //         const timeLeft = (expirationTime - now) / 1000
-  //
-  //         await respond({
-  //           embeds: [Utils.generateErrorMessage(`Пожалуйста, подождите еще ${timeLeft.toFixed(2)} секунд перед тем как использовать \`${realCommandName}\`!`,
-  //             ErrorMessageType.Warning)]
-  //         },
-  //         timeLeft * 1000 + 1000)
-  //         return
-  //       }
-  //     } else {
-  //       timestamps.set(member.id, now)
-  //       setTimeout(() => timestamps.delete(member.id), cooldownAmount)
-  //     }
-  //
-  //     await command?.execute({
-  //       guild,
-  //       user: member.user,
-  //       voice: member?.voice?.channel as VoiceBasedChannel,
-  //       text: channel,
-  //       client: this.client,
-  //       args,
-  //       respond,
-  //       send,
-  //       message,
-  //       meta
-  //     }).catch(err => logger.error({ err, ...meta }, 'Error executing command'))
   //   }
   // }
 }
